group faces based on plane only
no adjacency required.  This is because some faces have coincident
edges, which are nevertheless different edges.  This simpler approach is
consistent with the original qrad3 interpolation and seems to work well
here.  Also reduce the threshold for sampling points to be any face that
has non-zero area.  This might be noisy but works well until we come up
with another scheme for dealing with small faces